jetson_app/requirements.txt

pyserial
opencv-python
numpy
torch
torchvision
